requirements.txt
Django==2.2.16
djangorestframework==3.12.4
django-filter==2.4.0
pandas==1.4.3
openpyxl==3.0.10
psycopg2-binary==2.8.6